updated footer test
